Queryhandler successfully getting list of players
The records of players who scanned the same QR code are now being
retrived and displayed. All thats left to do is parse the names
out of these and make the activity look nice. Just merging it now since
I'm eager to get Queryhandler onto main so others can start adding to it.