Revised MainActivity
The queries related to reconstructing the Account and reconstructing the Records were not compatible. There was a sleugh of bugs and unintended behaviour.
This commit has them working together somewhat - but still faces issues wrt the database code being executed asynchronously to the code in the activities.
Currently my testing and commit has this: When it logs you in and sends you to AccountActivity it still has not loaded in your Records to display. If you visit a different activity then visit AccountActivity again, they are present.
Next commit: Use QueryHandler and QueryCallback to make this seemless.