Temporary query workarounds.
There are still some things wrong with the queries in MainActivity. The biggest offender is that the nested query to reconstruct account records happens asynchronously - so you can visit your profile before your records are available.
For now, after login it intents to MapActivity instead of AccountActivity, to give it time to finish reconstruction.